+++ .gitignore
@@ -11,3 +11,7 @@ mutants.out.old/
 /templates/release_scripts/enum.txt
 /templates/release_scripts/name.txt
 /templates/release_scripts/code.txt
+
+/templates/languages/enum.txt
+/templates/languages/name.txt
+/templates/languages/code.txt
