+++ templates/languages/file_template.txt
@@ -1,0 +1,53 @@
+/// The language the release title and track titles are written in. The possible values are taken
+/// from the [ISO 639-3](https://en.wikipedia.org/wiki/ISO_639-3) standard.
+///
+/// The values for this enum have been generated with the following command:
+///
+/// ```bash
+/// $ curl -s https://musicbrainz.org/statistics/languages-scripts | \
+///     grep -Eo '<td>[^<]*</td><td class="t"><a href="https://musicbrainz.org/search\?query=lang%3A%22[^"]*%22' | \
+///     sort | \
+///     sed -e 's,<td>\([^<]*\)</td><td class="t"><a href="https://musicbrainz.org/search?query=lang%3A%22\([^"]*\)%22,\/\/\/ \1\n\u\2\,,' -e "s/&#x27;/'/"
+/// ```
+#[non_exhaustive]
+#[derive(Debug, serde::Serialize, serde::Deserialize, PartialEq, Eq, Clone)]
+#[serde(rename_all = "lowercase")]
+pub enum Language {
+
+}
+
+impl Language {
+    /// Get the human-readable name used by MusicBrainz.
+    ///
+    /// Generated using:
+    /// ```console
+    /// $ curl -s https://musicbrainz.org/statistics/languages-scripts | \
+    ///     grep -Eo '<td>[^<]*</td><td class="t"><a href="https://musicbrainz.org/search\?query=lang%3A%22[^"]*%22' | \
+    ///     sort | \
+    ///     sed -e 's,<td>\([^<]*\)</td><td class="t"><a href="https://musicbrainz.org/search?query=lang%3A%22\([^"]*\)%22,Self::\2 => "\1"\,,' -e "s/&#x27;/'/"
+    /// ```
+    ///
+    /// and using editor features to fix capitalization of the variants.
+    pub fn name(&self) -> &'static str {
+        match &self {
+
+        }
+    }
+
+    /// Get the [ISO 639-3](https://en.wikipedia.org/wiki/ISO_639-3) code as [`str`].
+    ///
+    /// Generated using:
+    /// ```console
+    /// $ curl -s https://musicbrainz.org/statistics/languages-scripts | \
+    ///     grep -Eo '<td>[^<]*</td><td class="t"><a href="https://musicbrainz.org/search\?query=lang%3A%22[^"]*%22' | \
+    ///     sort | \
+    ///     sed 's,<td>\([^<]*\)</td><td class="t"><a href="https://musicbrainz.org/search?query=lang%3A%22\([^"]*\)%22,Self::\2 => "\2"\,,'
+    /// ```
+    ///
+    /// and using editor features to fix capitalization of the variants.
+    pub fn code(&self) -> &'static str {
+        match &self {
+
+        }
+    }
+}
